[REFACTOR] #60 - Notifications 화면에서 데이터 로직 이전
