+++ app/src/main/java/org/woowatechcamp/githubapplication/presentation/home/notifications/NotiViewModel.kt
@@ -37,55 +37,26 @@ class NotiViewModel @Inject constructor(
 
     fun getNoti() = viewModelScope.launch {
         _notiState.emit(UiState.Loading)
-        notiRepository.getNoti()
-            .onSuccess {
-                _notiState.emit(UiState.Success(it))
-                getComments(it)
+        with(notiRepository.getNoti()) {
+            _notiState.emit(this)
+            if (this is UiState.Success) {
+                getComments(this.value)
             }
-            .onFailure { e ->
-                _notiState.emit(UiState.Error(e.message ?: "알림을 가져오는 데 실패했습니다."))
-            }
+        }
     }
 
     private suspend fun getComments(notiList: List<NotiModel>) {
         notiList.forEach { item ->
-            notiRepository.getComment(item)
-                .onSuccess {
-                    val newItem = item.refreshComment(it.size)
-                    _notiCommentState.emit(UiState.Success(newItem))
-                }
-                .onFailure { e ->
-                    _notiCommentState.emit(
-                        UiState.Error(e.message ?: "Comment 개수를 가져오는 데 실패했습니다.")
-                    )
-                }
+            _notiCommentState.emit(
+                notiRepository.getComment(item)
+            )
         }
     }
 
     fun markNoti(threadId: String) = viewModelScope.launch {
         _markState.emit(UiState.Loading)
-        notiRepository.markNoti(threadId)
-            .onSuccess { res ->
-                _markState.emit(
-                    when (res.code()) {
-                        205 -> {
-                            UiState.Success("Success")
-                        }
-                        304 -> {
-                            UiState.Error("Not Modified")
-                        }
-                        403 -> {
-                            UiState.Error("Forbidden")
-                        }
-                        else -> {
-                            UiState.Error("알림 읽음을 처리하는 데 오류가 발생했습니다.")
-                        }
-                    }
-                )
-            }.onFailure { e ->
-                _markState.emit(
-                    UiState.Error(e.message ?: "알림 읽음을 처리하는 데 실패했습니다.")
-                )
-            }
+        _markState.emit(
+            notiRepository.markNoti(threadId)
+        )
     }
 }
